Move resistance stats to new file
