+++ src/attack_calculator.ts
@@ -1,4 +1,4 @@
-import { enumerateEnumValues } from 'utils';
+import { DamageType, ResistanceStats } from 'resistance_stats';
 
 // Note that these enums/ util classes should likely be moved elsewhere eventually
 export enum Attribute {
@@ -20,20 +20,6 @@ export enum EvasiveStatType {
     Fortitude,
     Reflex,
     Willpower,
-}
-
-export enum DamageType {
-    Slashing,
-    Bludgeoning,
-    Piercing,
-    Fire,
-    Water,
-    Air,
-    Earth,
-    Poison,
-    Radiant,
-    Necrotic,
-    Psychic,
 }
 
 export interface Weapon {
@@ -128,33 +114,6 @@ export class AttributeStats {
         }
 
         return Math.ceil(0.75 * statSum);
-    }
-}
-
-export interface ResistanceStat {
-    percent: number;
-    flat: number;
-}
-
-export class ResistanceStats {
-    damageTypeToResistance: Record<DamageType, ResistanceStat>;
-
-    constructor(damageTypeToNonZeroResistances?: Partial<Record<DamageType, ResistanceStat>>) {
-        // Forced cast is because filling the object dynamically plays poorly with Record
-        this.damageTypeToResistance = {} as Record<DamageType, ResistanceStat>;
-        for (const damageType of enumerateEnumValues<DamageType>(DamageType)) {
-            const nonZeroResistance = damageTypeToNonZeroResistances?.[damageType];
-            this.damageTypeToResistance[damageType] =
-                (nonZeroResistance === undefined) ? {percent: 0, flat: 0} : nonZeroResistance;
-        }
-    }
-
-    get(damageType: DamageType) : ResistanceStat {
-        return this.damageTypeToResistance[damageType];
-    }
-
-    set(damageType: DamageType, resistanceStat: ResistanceStat) {
-        this.damageTypeToResistance[damageType] = resistanceStat;
     }
 }
